add ability to remove trip, layout fixes, refactor, add two more resorts - Tucker

+++ dvc-calc-swiftui/Models/Trip.swift
@@ -20,12 +20,12 @@ class Trip: Comparable {
     var viewTypeId: UUID
     var checkInDate: Date
     var checkOutDate: Date
-    var points: Int16
+    var points: Int
     var borrowedFromLastYear: Int
     var borrowedFromNextYear: Int
     var contract: Contract? = nil
     
-    init(resortId: UUID, roomTypeId: UUID, viewTypeId: UUID, checkInDate: Date, checkOutDate: Date, points: Int16, borrowedFromLastYear: Int, borrowedFromNextYear: Int, contract: Contract?) {
+    init(resortId: UUID, roomTypeId: UUID, viewTypeId: UUID, checkInDate: Date, checkOutDate: Date, points: Int, borrowedFromLastYear: Int, borrowedFromNextYear: Int, contract: Contract?) {
         self.resortId = resortId
         self.roomTypeId = roomTypeId
         self.viewTypeId = viewTypeId
